refactor: consolidate category configs into shared file (#1312)
